Move all test cases from `MultiplayerFlowTests` to their classes
